+++ examples/conditional_routing/src/main.rs
@@ -1,0 +1,56 @@
+use std::time::Duration;
+
+use picoserve::routing::get;
+
+#[tokio::main(flavor = "current_thread")]
+async fn main() -> anyhow::Result<()> {
+    let port = 8000;
+
+    let common_routes = picoserve::Router::new().route("/", get(|| async { "Hello World" }));
+
+    // If you change this to false, `http://localhost:8000/other` will return a 404 instead of "Other Route!".
+    let include_other_route = true;
+
+    let app = std::rc::Rc::new(if include_other_route {
+        common_routes
+            .route("/other", get(|| async { "Other Route!" }))
+            .either_left_route()
+    } else {
+        common_routes.either_right_route()
+    });
+
+    let config = picoserve::Config::new(picoserve::Timeouts {
+        start_read_request: Some(Duration::from_secs(5)),
+        read_request: Some(Duration::from_secs(1)),
+        write: Some(Duration::from_secs(1)),
+    })
+    .keep_connection_alive();
+
+    let socket = tokio::net::TcpListener::bind((std::net::Ipv4Addr::LOCALHOST, port)).await?;
+
+    println!("http://localhost:{port}/");
+
+    tokio::task::LocalSet::new()
+        .run_until(async {
+            loop {
+                let (stream, remote_address) = socket.accept().await?;
+
+                println!("Connection from {remote_address}");
+
+                let app = app.clone();
+                let config = config.clone();
+
+                tokio::task::spawn_local(async move {
+                    match picoserve::serve(&app, &config, &mut [0; 2048], stream).await {
+                        Ok(handled_requests_count) => {
+                            println!(
+                                "{handled_requests_count} requests handled from {remote_address}"
+                            )
+                        }
+                        Err(err) => println!("{err:?}"),
+                    }
+                });
+            }
+        })
+        .await
+}